Pull r64124 into Dib:
Turns out AnalyzeBranch can modify the mbb being analyzed. This is a nasty
suprise to some callers, e.g. register coalescer. For now, add an parameter that
tells AnalyzeBranch whether it's safe to modify the mbb. A better solution is
out there, but I don't have time to deal with it right now.


git-svn-id: https://llvm.org/svn/llvm-project@64133 91177308-0d34-0410-b5e6-96231b3b80d8