Pull 65497 and 65498 into Dib:
If an available register falls through to a succ block, unset the last
kill. Sorry, it's impossible to reduce a sensible test case. It basically
requires the moon and stars to align in order to cause a failure.

The last commit was overly conservative. It's ok to reuse value that's already
marked livein.


git-svn-id: https://llvm.org/svn/llvm-project@65515 91177308-0d34-0410-b5e6-96231b3b80d8