Completely rewrite the JnjvmModule abstraction in the JVM. We now have two separate kinds of modules: AOT and JIT. The JVM lets the tool decide which JIT it wants.
Because I had to use virtual functions, JnjvmModule could not extend
llvm::Module anymore. And since there was no real reason for it to
extend the llvm::Module class, I moved the llvm::Module object to
a field.

Also as a side-effect, I had to fix stack unwinding problems. Functions
that are involved in Java/native cross-calls and functions that get
the frame pointer now have a noinline attribute. This fixes the Debug
build on x86.



git-svn-id: https://llvm.org/svn/llvm-project@65555 91177308-0d34-0410-b5e6-96231b3b80d8