--- Merging (from foreign repository) r67134 into '.': A    test/CodeGen/Generic/2009-03-17-LSR-APInt.ll U    lib/Transforms/Scalar/LoopStrengthReduce.cpp
LSR shouldn't ever try to hack on integer IV's larger than 64-bits.  Right now
it is not APInt clean, but even when it is it needs to be evaluated carefully
to determine whether it is actually profitable.

This fixes a crash on PR3806

--- Merging (from foreign repository) r67139 into '.':
A    test/Transforms/DeadArgElim/2009-03-17-MRE-Invoke.ll
U    lib/Transforms/IPO/DeadArgumentElimination.cpp
Skipped missing target: 'test/CodeGen/X86/call-imm.ll'

Fix PR3807 by inserting 'insertelement' instructions in the normal dest of
an invoke instead of after the invoke (in its block), which is invalid.

--- Merging (from foreign repository) r67142 into '.':
U    lib/Target/X86/X86Instr64bit.td

Disable the "call to immediate" optimization on x86-64.  It is
not safe in general because the immediate could be an arbitrary
value that does not fit in a 32-bit pcrel displacement.
Conservatively fall back to loading the value into a register
and calling through it.

We still do the optzn on X86-32.


git-svn-id: https://llvm.org/svn/llvm-project@67144 91177308-0d34-0410-b5e6-96231b3b80d8