--- Merging (from foreign repository) r68598 into '.': A    test/Transforms/InstCombine/2009-04-07-MulPromoteToI96.ll U    lib/Transforms/Scalar/InstructionCombining.cpp
Instcombine should not promote whole computation trees to "strange"
integer types, unless they are already strange.  This prevents it from
turning the code produced by SROA into crazy libcalls and stuff that 
the code generator can't handle.  In the attached example, the result
was an i96 multiply that caused the x86 backend to assert.

Note that if TargetData had an idea of what the legal types are for
a target that this could be used to stop instcombine from introducing
i64 muls, as Scott wanted.


git-svn-id: https://llvm.org/svn/llvm-project@68622 91177308-0d34-0410-b5e6-96231b3b80d8