--- Merging (from foreign repository) r68576 into '.': U    test/CodeGen/X86/subreg-to-reg-1.ll U    test/CodeGen/X86/iv-users-in-other-loops.ll A    test/CodeGen/X86/subreg-to-reg-3.ll A    test/CodeGen/X86/subreg-to-reg-4.ll A    test/CodeGen/X86/subreg-to-reg-5.ll A    test/CodeGen/X86/subreg-to-reg-6.ll U    test/CodeGen/X86/twoaddr-coalesce.ll U    test/CodeGen/X86/byval7.ll D    test/CodeGen/X86/2008-05-06-SpillerBug.ll U    include/llvm/Target/TargetLowering.h U    lib/CodeGen/LiveInterval.cpp U    lib/CodeGen/LiveIntervalAnalysis.cpp U    lib/CodeGen/SimpleRegisterCoalescing.cpp U    lib/CodeGen/TwoAddressInstructionPass.cpp U    lib/CodeGen/SelectionDAG/DAGCombiner.cpp U    lib/CodeGen/SelectionDAG/TargetLowering.cpp U    lib/Target/X86/X86Instr64bit.td U    lib/Target/X86/X86ISelLowering.cpp U    lib/Target/X86/X86ISelLowering.h
Implement support for using modeling implicit-zero-extension on x86-64
with SUBREG_TO_REG, teach SimpleRegisterCoalescing to coalesce
SUBREG_TO_REG instructions (which are similar to INSERT_SUBREG
instructions), and teach the DAGCombiner to take advantage of this on
targets which support it. This eliminates many redundant
zero-extension operations on x86-64.

This adds a new TargetLowering hook, isZExtFree. It's similar to
isTruncateFree, except it only applies to actual definitions, and not
no-op truncates which may not zero the high bits.

Also, this adds a new optimization to SimplifyDemandedBits: transform
operations like x+y into (zext (add (trunc x), (trunc y))) on targets
where all the casts are no-ops. In contexts where the high part of the
add is explicitly masked off, this allows the mask operation to be
eliminated. Fix the DAGCombiner to avoid undoing these transformations
to eliminate casts on targets where the casts are no-ops.

Also, this adds a new two-address lowering heuristic. Since
two-address lowering runs before coalescing, it helps to be able to
look through copies when deciding whether commuting and/or
three-address conversion are profitable.

Also, fix a bug in LiveInterval::MergeInClobberRanges. It didn't handle
the case that a clobber range extended both before and beyond an
existing live range. In that case, multiple live ranges need to be
added. This was exposed by the new subreg coalescing code.

Remove 2008-05-06-SpillerBug.ll. It was bugpoint-reduced, and the
spiller behavior it was looking for no longer occurrs with the new
instruction selection.

--- Merging (from foreign repository) r68579 into '.':
U    test/CodeGen/X86/dyn-stackalloc.ll

Update this test for recent codegen improvements. CodeGen is now
using an lea in place of a mov and an add for this test.

--- Merging (from foreign repository) r68580 into '.':
G    test/CodeGen/X86/dyn-stackalloc.ll

Fully escape the grep string for this test.


git-svn-id: https://llvm.org/svn/llvm-project@68626 91177308-0d34-0410-b5e6-96231b3b80d8