--- Merging (from foreign repository) r69743 into '.': A    test/CodeGen/X86/2009-04-21-NoReloadImpDef.ll U    include/llvm/CodeGen/LiveIntervalAnalysis.h U    lib/CodeGen/LiveIntervalAnalysis.cpp U    lib/CodeGen/RegAllocPBQP.cpp U    lib/CodeGen/Spiller.cpp U    lib/CodeGen/Spiller.h U    lib/CodeGen/RegAllocLinearScan.cpp
It has finally happened. Spiller is now using live interval info.

This fixes a very subtle bug. vr defined by an implicit_def is allowed overlap
with any register since it doesn't actually modify anything. However, if it's
used as a two-address use, its live range can be extended and it can be
spilled. The spiller must take care not to emit a reload for the vn number
that's defined by the implicit_def. This is both a correctness and performance
issue.


git-svn-id: https://llvm.org/svn/llvm-project@69746 91177308-0d34-0410-b5e6-96231b3b80d8