--- Merging r71224 into '.': A    test/Transforms/ScalarRepl/2009-05-08-I1Crash.ll U    lib/Transforms/Scalar/ScalarReplAggregates.cpp
fix RewriteStoreUserOfWholeAlloca to use the correct type size
method, fixing a crash on PR4146.  While the store will
ultimately overwrite the "padded size" number of bits in memory,
the stored value may be a subset of this size.  This function
only wants to handle the case where all bits are stored.


git-svn-id: https://llvm.org/svn/llvm-project@71229 91177308-0d34-0410-b5e6-96231b3b80d8