--- Merging r73730 into 'gcc/llvm-convert.cpp': U    gcc/llvm-convert.cpp
fix rdar://6983634 - when inserting tail padding to round up an initializer to
the right size, we have to be careful when the max struct alignment will cause the
size-with-alignment to be too big.  In this case, switch the struct to being packed.

This fixes rdar://6983634 and test/FrontendC/2009-06-18-StaticInitTailPadPack.c


git-svn-id: https://llvm.org/svn/llvm-project@75694 91177308-0d34-0410-b5e6-96231b3b80d8